+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Deprecated
 ### Removed
 ### Fixed
+- Fixed handling of leading newlines in `pre` and `textarea` elements
 ### Security
 
 ## [0.2.1] - 2019-03-27